model_deploy/requirements.txt
django == 2.2.7
numpy == 1.18.1 
djangorestframework == 3.8.2
gunicorn == 20.0.4 
sklearn == 0.0